# Improve token search performance