# Description

- Removed extraneous loading prop
- First keypress on search input caused UI to freeze for ~3 seconds. Removed nested loops from `mapTokenListTokens`.

## Type of change

- [x] Code refactor / cleanup

## How should this be tested?

Please provide instructions so we can test. Please also list any relevant details for your test configuration.

1. Go to home page
2. Click "Filter by token" button
3. Start typing to token search
4. Token list should update immediately

## Visual Context

### First keypress event run time dropped from 3.38s to 43ms
| Before | Now |
| ------ | ---- |
| <img width="640" alt="Screen Shot 2022-07-01 at 12 39 44" src="https://user-images.githubusercontent.com/28311328/176871207-48c30ad9-14f9-4434-8187-4c62871fe40e.png">   | <img width="640" alt="Screen Shot 2022-07-01 at 12 37 41" src="https://user-images.githubusercontent.com/28311328/176871265-b7017e02-cf89-47d8-bf56-3472385b6616.png">  |


## Checklist:

- [x] I have performed a self-review of my own code
- [x] I have requested at least 2 reviews (If the PR is significant enough, use best judgement here)
- [x] I have commented my code where relevant, particularly in hard-to-understand areas
- [x] If package-lock.json has changes, it was intentional.
- [x] The base of this PR is `master` if hotfix, `develop` if not